Rename event config to handler config
